tests-python/requirements.txt
allure-behave
allure-pytest
behave
playwright
pytest
pytest-playwright
requests
#selenium
